requirements.txt
asgiref==3.8.1
Django==4.2.18
resend==2.13.1
django-admin-interface==0.29.4
django-colorfield==0.12.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gunicorn==23.0.0
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.0
python-slugify==8.0.4
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.12.2
requests==2.32.3
django-filter==24.1
boto3==1.28.38
django-storages==1.14.2
whitenoise==6.9.0
xlsxwriter==3.2.5
qrcode==7.4.2
beautifulsoup4==4.12.3